+++ packages/@expo/cli/src/export/fork-bundleAsync.ts
@@ -22,6 +22,7 @@ import {
   ExpoMetroBundleOptions,
   getMetroDirectBundleOptionsForExpoConfig,
 } from '../start/server/middleware/metroOptions';
+import { CommandError } from '../utils/errors';
 
 export type MetroDevServerOptions = LoadOptions;
 
@@ -106,6 +107,17 @@ export async function createBundlesAsync(
   );
 }
 
+function assertMetroConfig(
+  config: ConfigT
+): asserts config is ConfigT & { serializer: NonNullable<ConfigT['serializer']> } {
+  if (!config.serializer?.customSerializer) {
+    throw new CommandError(
+      'METRO_CONFIG_MALFORMED',
+      `The Metro bundler configuration is missing required features from 'expo/metro-config' and cannot be used with Expo CLI. Ensure the metro.config.js file is extending 'expo/metro-config'. Learn more: https://docs.expo.dev/guides/customizing-metro`
+    );
+  }
+}
+
 async function bundleProductionMetroClientAsync(
   projectRoot: string,
   expoConfig: ExpoConfig,
@@ -122,6 +134,8 @@ async function bundleProductionMetroClientAsync(
     exp: expoConfig,
     isExporting: true,
   });
+
+  assertMetroConfig(config);
 
   const metroServer = await Metro.runMetro(config, {
     watch: false,
@@ -310,6 +324,8 @@ async function forkMetroBuildAsync(
     // Custom options we pass to the serializer to emulate the URL query parameters.
     serializerOptions: options.serializerOptions,
   };
+
+  assertMetroConfig(metro._config);
 
   const bundle = await metro._config.serializer.customSerializer!(
     entryPoint,